Updated README.md
I added a summary to explain what SMD does, then created a table of content that shows new users what the project will cover, sending them to key points if they click on the contents. I then added a box for the commands you can run, then grouped together the information according to the table of contents. Lastly, grouped together the pictures of each GUI for representation of what the user will see on different platforms.